chore(controllers): de-export helpers, fix lint, split network.controller experiment routes (#1031)
* chore(controllers): de-export internal-only helpers and fix stale spec lint

- Narrow parseApiKeyMetadata (mcp), hasAtLeastOneSocial/shouldAutoGenerateProfile
  (auth) to module-private; each was exported but referenced only in-file.
- Remove unused start/duration vars in enrichment.controller.spec (pre-existing
  eslint no-unused-vars warning).

* refactor(controllers): extract experiment routes into NetworkExperimentController

Split the 1037-LOC network.controller.ts (31 routes) by moving the 8
experiment-network routes — headless master-key signup/lookup, CSV import
parse/apply, single-email invite/resend, and master-key/network-key rotation —
into a sibling NetworkExperimentController. Both share @Controller('/networks');
the router dispatches by full method+path (as NetworkOpportunityController
already does), so no route paths change.

The new controller exclusively owns experimentService, networkInvitationService,
ExperimentMasterKeyGuard, EMAIL_REGEX, and the assertExperimentOwner/parseCsv*
private helpers, which all move out wholesale — NetworkController drops to 582
LOC of core CRUD/membership. Registered in main.ts beside NetworkController.

Tests: retarget inviteMember/rotateMasterKey calls in network.controller.spec
to the new controller instance. 42/42 network+invitation specs pass; tsc and
eslint clean.
