+++ backend/src/main.ts
@@ -8,6 +8,7 @@ import { ToolController } from './controllers/tool.controller';
 import { ToolService } from './services/tool.service';
 import { S3StorageAdapter } from './adapters/storage.adapter';
 import { NetworkController } from './controllers/network.controller';
+import { NetworkExperimentController } from './controllers/network-experiment.controller';
 import { IntentController } from './controllers/intent.controller';
 import { LinkController } from './controllers/link.controller';
 import { OpportunityController, NetworkOpportunityController } from './controllers/opportunity.controller';
@@ -364,6 +365,7 @@ controllerInstances.set(AuthController, new AuthController());
 controllerInstances.set(EnrichmentController, new EnrichmentController());
 controllerInstances.set(ChatController, new ChatController());
 controllerInstances.set(NetworkController, new NetworkController());
+controllerInstances.set(NetworkExperimentController, new NetworkExperimentController());
 controllerInstances.set(IntentController, new IntentController());
 controllerInstances.set(LinkController, new LinkController());
 controllerInstances.set(OpportunityController, new OpportunityController());
